# improve event performance