Hey. As I digged into dEvents, i endet with these simple improvements.

**simple item drop:**
_without: 0.21%	   1.91%	  0.01 s	     0.95 ms
with: 0.03%	   0.35%	  0.00 s	     0.18 ms_

**written book drop:**
_without: 1.02%	  58.65%	  0.03 s	    29.32 ms
with: 0.16%	   1.12%	  0.01 s	     0.56 ms_

**placing of stone:**
_without: 0.29%	   2.35%	  0.02 s	     1.17 ms
with: 0.05%	   0.48%	  0.00 s	     0.24 ms_

So, for default items, it can speed up events by 100-400% depending on items and events.

The idea is again, to skip all further checks, when the Material reference doesn't match.

It should affect all events, as they use tryItem/tryMaterial.
(i also skipped this null check in PlayerClicksBlock, as tryItem already checks for null)

Yet i haven't found negative side effects of the changes, you may take still a closer look.