@mcmonkey4eva 2018-06-22
`i haven't found negative side effects of the changes`

How about `on player right clocks block:` which will not work with that code. Gotta be more careful than that editing high-usage code.